Finish renaming distributed simulator -> simulator
