+++ dist_ir/executor/__init__.py
@@ -1,4 +1,4 @@
-from .distributed_simulator import Simulator
+from .simulator import Simulator
 from .sequential_executor import SequentialExecutor
 from .type_inference import infer_types
 from .absint import AbstractInterpreter, AbstractState
